modules/ObjectDetectionService/requirements.txt
argparse
numpy<2  # 2.0 and above breaks compatibility with cv2
opencv-python~=4.5.3.56
tflite-support==0.4.2
protobuf>=3.18.0,<4
# Added
# flask==2.2.2 
# pillow==7.1.2
flask
pillow